[#190] refactor: EmulToken 코드 리뷰 반영
- 폴더명 수정
- Repository -> Component
