+++ rest/src/main/java/com/wherecar/rest/emulauth/hubtmp/token/infrastructure/EmulTokenReaderImpl.java
@@ -2,9 +2,9 @@ package com.wherecar.rest.emulauth.hubtmp.token.infrastructure;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.stereotype.Repository;
+import org.springframework.stereotype.Component;
 
-@Repository
+@Component
 @RequiredArgsConstructor
 public class EmulTokenReaderImpl implements EmulTokenReader {
